+++ src/imio/history/safe_utils.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from imio.history.utils import getLastAction
+from imio.history.utils import getLastWFAction
+from imio.history.utils import getPreviousEvent